refactor: extract agent-runner.ts and run-modes.ts from run.ts (#447)
Steps 6-7: Extract runAgent() into agent-runner.ts and all squad execution
modes (autopilot, lead, squad loop, post-evaluation) into run-modes.ts.

run.ts: 2930 → 320 lines (89% reduction). Now contains only command
registration, runCommand() routing, and runSquad() dispatch.
